fastapi-docker/requirements.txt
fastapi
uvicorn
transformers
Pillow
torch
python-multipart